+++ DataTransferObject/Domain/Master/MBde.cs
@@ -11,26 +11,26 @@ namespace DataTransferObject.Domain.Master
     public class MBde :Common
     {
         [Key]
-        public short BdeId { get; set; }
+        public byte BdeId { get; set; }
         [Required(ErrorMessage = "required!")]
         [RegularExpression("^[a-zA-Z0-9 ]*$", ErrorMessage = "Only Alphabets and Numbers allowed.")]
         public string BdeName { get; set; }
 
 
         [ForeignKey("Comd")]
-        public short ComdId { get; set; }
+        public byte ComdId { get; set; }
 
         public MComd? Comd { get; set; }
 
         [ForeignKey("MCorps")]
-        public short CorpsId { get; set; }
+        public byte CorpsId { get; set; }
         public MCorps? Corps { get; set; }
 
         [ForeignKey("MDiv")]
 
         [Required(ErrorMessage = "required!")]
 
-        public short DivId { get; set; }
+        public byte DivId { get; set; }
         public MDiv? Div { get; set; }   
 
     }
