+++ DataTransferObject/Domain/Master/MRank.cs
@@ -13,7 +13,7 @@ namespace DataTransferObject.Domain.Master
     { 
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public short RankId { get; set; }
+        public byte RankId { get; set; }
 
         [Required(ErrorMessage = "required!")]
         [RegularExpression("^[a-zA-Z0-9 ]*$", ErrorMessage = "Only Alphabets and Numbers allowed.")]
@@ -22,8 +22,8 @@ namespace DataTransferObject.Domain.Master
         [RegularExpression("^[a-zA-Z0-9 ]*$", ErrorMessage = "Only Alphabets and Numbers allowed.")]
         public string RankAbbreviation { get; set; }
 
-        public short Orderby { get; set; }
+        public byte Orderby { get; set; }
         [Required(ErrorMessage = "required!")]
-        public short Type { get; set; }
+        public byte Type { get; set; }
     }
 }
